게시물 조회 추가 - 게시물 업로드 시간, 수정시간 칼럼 추가필요 - 모집인원 관리 db table 따로 생성할 것인지 상의필요

+++ src/main/java/com/jhcompany/studyLink/recruit/repository/RecruitPostRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RecruitPostRepository extends JpaRepository<RecruitPostEntity, Integer> {
+public interface RecruitPostRepository extends JpaRepository<RecruitPostEntity, Long> {
     RecruitPostEntity findByCid(String userId);
 
 //    @Query("SELECT count(r) FROM RecruitPostEntity r WHERE r.cid = :cid")
